use float64 by default

+++ src/eko/_parameters.py
@@ -1,4 +1,4 @@
 import numpy as np
 
-t_float = np.float32
+t_float = np.float64
 t_complex = np.complex64
